+++ pages/index.vue
@@ -2,26 +2,45 @@
   <v-app id="inspire">
     <Header :is-login="isLogin"></Header>
     <v-main class="grey lighten-2">
-      <v-container>
-        <Dialog 
-          v-if="dialog"
-          :dialog="dialog"
-          :status="dialogTaskStatus"
-          @closeDialog="closeDialog"
-          @addTask="addTask"
-        ></Dialog>
-        <v-row v-if="taskStatusList">
-          <task-sheet
-            v-for="(item, index) in taskStatusList"
-            :key="index"
-            :status="item.status"
-            :color="item.color"
-            :tasks="filteredTasks(item.status)"
-            @openDialog="openDialog"
-            @deleteTask="deleteTask"
-          ></task-sheet>
-        </v-row>
-      </v-container>
+      <v-card
+        class="mx-auto mt-10 px-10 py-10"
+        max-width="500"
+      >
+        <v-form
+          ref="form"
+          lazy-validation
+        >
+          <v-card-title class="d-flex justify-center font-weight-bold">ログイン</v-card-title>
+
+          <v-text-field
+            v-model="email"
+            :rules="emailRules"
+            label="メールアドレス"
+            required
+          ></v-text-field>
+
+          <v-text-field
+            v-model="password"
+            :append-icon="show ? 'mdi-eye' : 'mdi-eye-off'"
+            :type="show ? 'text' : 'password'"
+            :rules="passwordRules"
+            label="パスワード"
+            required
+            @click:append="show = !show"
+          ></v-text-field>
+
+          <div class="d-flex justify-center mt-5">
+            <v-btn
+              color="blue"
+              dark
+              large
+              @click="login"
+            >
+              ログイン
+            </v-btn>
+          </div>
+        </v-form>
+      </v-card>
     </v-main>
   </v-app>
 </template>
@@ -29,86 +48,40 @@
 <script lang="ts">
 import Vue from 'vue'
 import Header from '../components/Header.vue'
-import TaskSheet from '../components/TaskSheet.vue'
-import Dialog from '../components/Dialog.vue'
-import type { CreateTaskRequest } from '$/types'
 
 export default Vue.extend({
   components: {
-    Header,
-    TaskSheet,
-    Dialog
+    Header
   },
   data() {
     return {
-      isLogin: true,
-      dialog: false,
-      dialogTaskStatus: '',
-      taskStatusList: [
-        {
-          status: 'TODO',
-          color: 'red'
-        },
-        {
-          status: 'DOING',
-          color: 'green'
-        },
-        {
-          status: 'DONE',
-          color: 'black'
-        }
+      isLogin: false,
+      email: '',
+      emailRules: [
+        (v: string) => !!v || 'メールアドレスは必須です！',
+        (v: string) => /.+@.+\..+/.test(v) || 'メールアドレスの形式が正しくありません！',
       ],
-      tasks: [
-        {
-          id: 0,
-          title: '',
-          status: ''
-        }
+      password: '',
+      show: false,
+      passwordRules: [
+        (v: string) => !!v || 'パスワードは必須です！',
+        (v: string) => v.length >= 8 || 'パスワードは8文字以上にしてください！',
       ]
-    };
-  },
-  async mounted() {
-    this.tasks = await this.$api.tasks.$get({
-      headers: {
-        Authorization: `Bearer ${this.$store.getters.token}`
-      }
-    })
+    }
   },
   methods: {
-    filteredTasks(status: string) {
-      return this.tasks.filter(task => task.status === status)
-    },
-    async addTask(task: CreateTaskRequest) {
-      this.closeDialog()
+    async login () {
+      if(!(this.$refs.form as Vue & { validate: () => boolean }).validate()) return
 
-      const newTask = await this.$api.tasks.$post({
-        headers: {
-          Authorization: `Bearer ${this.$store.getters.token}`
-        },
-        body: task
-      })
+      try {
+        await this.$store.dispatch('login', { email: this.email, password: this.password })
+      } catch (e) {
+        if(e.response.status === 404)
+          window.alert("該当するユーザが存在しません。\nメールアドレスまたはパスワードに間違いがないか確認してもう一度お試しください！")
+      }
 
-      this.tasks.push({
-        id: newTask.taskId,
-        ...task
-      })
-    },
-    openDialog(status: string) {
-      this.dialogTaskStatus = status
-      this.dialog = true
-    },
-    closeDialog() {
-      this.dialog = false
-    },
-    async deleteTask(id: number) {
-      await this.$api.tasks._taskId(id).$delete({
-        headers: {
-          Authorization: `Bearer ${this.$store.getters.token}`
-        }
-      })
-
-      this.tasks = this.tasks.filter(task => task.id !== id)
+      this.$router.push('/tasks')
     }
-  }
+  },
 })
 </script>
